refactor:Combine DecryptData and EncryptData into a single file to avoid having similar type models scattered around
