+++ crypto_py/AES/_aes.py
@@ -15,8 +15,7 @@
 from Cryptodome.Cipher import AES as aes
 
 from ..models._options import Options
-from ..models._decrypt_data import DecryptData
-from ..models._encrypt_data import EncryptData
+from ..models._data_models import AESDecryptData,AESEncryptData
 from ..enc.enc import Utf8
 
 
@@ -24,7 +23,7 @@ from ..enc.enc import Utf8
 
 class AES:
     @staticmethod
-    def encrypt(data: str, key: bytes, options: Options|dict) -> EncryptData:
+    def encrypt(data: str, key: bytes, options: Options|dict) -> AESEncryptData:
 
         data = Utf8.parse(data)
         if isinstance(options,dict):
@@ -35,21 +34,21 @@ class AES:
                 padded_data = options.padding.pad(data)
                 # padded_data = data.encode('utf-8') + (16 - len(data.encode('utf-8')) % 16) * b'\0'
                 encrypted_data = cipher.encrypt(padded_data)
-                return EncryptData(encrypted_data)
+                return AESEncryptData(encrypted_data)
             case aes.MODE_CBC:
                 cipher = aes.new(key=key, mode=options.mode, iv=options.iv)
                 padded_data = options.padding.pad(data)
                 # padded_data = data.encode('utf-8') + (16 - len(data.encode('utf-8')) % 16) * b'\0'
                 encrypted_data = cipher.encrypt(padded_data)
-                return EncryptData(encrypted_data)
+                return AESEncryptData(encrypted_data)
 
     @staticmethod
-    def decrypt(data: str, key: bytes, options: Options) -> DecryptData:
+    def decrypt(data: str, key: bytes, options: Options) -> AESDecryptData:
         match options.mode:
             case aes.MODE_ECB:
                 cipher = aes.new(key, options.mode)
-                return DecryptData(cipher, data, options.padding)
+                return AESDecryptData(cipher, data, options.padding)
             case aes.MODE_CBC:
                 cipher = aes.new(key, options.mode, iv=options.iv)
-                return DecryptData(cipher, data, options.padding)
+                return AESDecryptData(cipher, data, options.padding)
