+++ crypto_py/models/_data_models.py
@@ -3,15 +3,15 @@
 
 # @IDE：PyCharm
 # @Python：3.1x
-# @Project：DP
+# @Project：crypto-js-to-py
 
 # -------------------------------
 
-# @fileName：_decrypt_data.py
-# @createTime：2024/7/3 9:56
+# @fileName：_data_models.py
+# @createTime：2024/7/12 17:16
 # @author：Primice
 
-# -------------------------------
+
 from Cryptodome.Cipher import AES
 from typing import Type
 import base64
@@ -20,8 +20,7 @@ import binascii
 from ..enc.enc import Utf8
 from ..pad.pad import Pad
 
-
-class DecryptData:
+class AESDecryptData:
     """ AES的解密数据对象。
 
     实例化后的类可以通过decode()方法获取utf8解密内容
@@ -61,4 +60,24 @@ class DecryptData:
         return self.__padding.pad(decrypted_data)
 
     def __str__(self) -> str:
-        return f"<class DecryptData data:'{self._data[:10]}'>"
+        return f"<class AESDecryptData data:'{self._data[:10]}'>"
+
+
+class AESEncryptData:
+    """AES密文的对象
+
+    因为实例化时传进来的值已经是密文，因此decode()和ciphertext只做编码过程
+
+    """
+    def __init__(self, encrypted_data):
+        self.encrypted_data = encrypted_data
+
+    def decode(self) -> str:
+        return str(base64.encodebytes(self.encrypted_data), encoding='utf-8').replace('\n', '')
+
+    @property
+    def ciphertext(self) -> bytes:
+        return binascii.hexlify(self.encrypted_data)
+
+    def __str__(self) -> str:
+        return f"<class EncryptData data:'{self.encrypted_data}'>"
